Modelling & Simple Texturing
Started creating a simple texture for an Amazonian worker, want to rework proportions and potentially rework UV layout now that I know more about how they work.

Created an Olympian model, which is a taller and buff version of the character model intended to be used as a boss. Rig was resized with it. Created and rigged a Centipede Bot middle segment.