Fixing Shooting (Didn't Succeed), Grapple Anim
Working on fixing shooting by setting the parent of the bullet to the owner in script, but it may just be going to the Player. Could set up an extension method to crawl upwards and find the true root.

Played with the animation tree to get the grapple animations to fire while holding the weapon still and trigger a function at the right point, then continue the animation when a hit is found.