Broke Character Body, Limb Signals, Grapple
Made a grappling hook projectile that fires on an event in the grapple animation and gets attached to a surface, haven't made it apply physics yet, thinking of just using a Signal with the Player Arms, could potentially collapse the arms scene into the player scene to streamline, mayhaps.

Experimented with getting characters to recoil when hit by a projectile, instead ran into issues with the scaling of the character model causing limbs to (wonderfully) inflate. Scaled the model in Blender to fix, but now need to rebuild the Body scene.