Fixed Animations, Body Rebuild, Bone Profile
Figured out how to use the Graph Editor to scale down all the keyframe data to match the new character model keyframes. Rebuild the physical bones for the body and included a RESET animation to improve stability. Created a humanoid skeleton profile for animation retargeting onto the Olympian.

Did a bunch of testing with the physical bones to try and get them to detach. and understand functionality.