Spider Walking V1
First version of the Spider Bot's procedural animation working, its a little jank and still requires a lot of tweaking (especially since it needs to move so slowly) but it works! Made a test scene with a different controller just for it as couldn't see it clearly in the main scene.